rebased against master; fixed a couple of small issues based on PR comments
