+++ src/bundler.d/development.rb
@@ -28,6 +28,4 @@ group :development do
   end
 
   gem 'factory_girl_rails', "~> 1.4.0"
-
-  gem "mocha", :require => false
 end
